+++ proto/eth/v1/attestation.proto
@@ -65,7 +65,7 @@ message AttestationData {
     // The committee index that submitted this attestation.
     uint64 index = 2 [(ethereum.eth.ext.cast_type) = "github.com/prysmaticlabs/prysm/v3/consensus-types/primitives.CommitteeIndex"];
 
-    // 32 byte root of the LMD GHOST block vote.
+    // 32 byte root of the LMD GHOST block vote.This
     bytes beacon_block_root = 3 [(ethereum.eth.ext.ssz_size) = "32"];
 
     // The most recent justified checkpoint in the beacon state
